Sketch of the changes required by #11
This is a starting point for the update required for the new version of
Pandoc.

The full update can't be done until there is a binary to run against, so
it will stay as a sketch until release date, but in the mean time this
gives us an idea of what will change (not much)

Also,

- Using globals() is probably very slow
- The from_json() function is quite cumbersome, perhaps we can simplify
it or split it in two? (check how it is called..)