Fix #28 - More robust convert_text()
Note that the API of convert_text() has changed

Before we were overloading the 'json' as input and output format, when
we in fact meant the AST generated by Panflute.

Now, we have all the Pandoc formats including json, plus a format
`panflute` that will create panflute elements.

Also, there is a `standalone` option (defaults to `False`) in case we
want a standalone document. We could have achieved this with the
`extra_args` option, but this also works with the `panflute` format:

`output_format='panflute', standalone=True` now will wrap up the
resulting panflute elements in a Doc() object.