Convert more builtins to MetaValues
This commit allows `builtin2meta` to convert `str` to `MetaSting` as
well as recursively converting `list`s and `dict`s.  The original
implementation handled conversion of native `float`, `bool`, and `int`
variables, but did not consider `str`.  On the container side, `list`
and `dict` variables simply passed the item through to `MetaList` and
`MetaMap` respectively.  This tacit assumed the contents was already a
`MetaValue` meaning constructs such as

    doc.metadata["references"] = doc.get_metadata("references")

would fail with a `TypeError`.